app/requirements.txt
flask
mysql-connector-python>=8.0.0  # Ensure pooling support